requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
autopep8==2.3.1
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nbqa==1.9.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
replicate==1.0.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tokenize_rt==6.1.0
tomli==2.0.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.0.6
